% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lexruntimeservice_package.R
\docType{package}
\name{paws.lexruntimeservice-package}
\alias{paws.lexruntimeservice}
\alias{paws.lexruntimeservice-package}
\title{paws.lexruntimeservice: Amazon Lex Runtime Service}
\description{
Amazon Lex provides both build and runtime
endpoints. Each endpoint provides a set of operations (API). Your
conversational bot uses the runtime API to understand user utterances
(user input text or voice). For example, suppose a user says \"I want
pizza\", your bot sends this input to Amazon Lex using the runtime
API. Amazon Lex recognizes that the user request is for the OrderPizza
intent (one of the intents defined in the bot). Then Amazon Lex
engages in user conversation on behalf of the bot to elicit required
information (slot values, such as pizza size and crust type), and then
performs fulfillment activity (that you configured when you created
the bot). You use the build-time API to create and manage your Amazon
Lex bot. For a list of build-time operations, see the build-time API,.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
